+++ NINA.Photon.Plugin.ASA/Trigger/MLPTafterTime.cs
@@ -57,7 +57,7 @@ namespace NINA.Photon.Plugin.ASA.MLTP
     [ExportMetadata("Category", "ASA Tools")]
     [Export(typeof(ISequenceTrigger))]
     [JsonObject(MemberSerialization.OptIn)]
-    public class MLPTafterTime : SequenceTrigger, IValidatable
+    public class MLPTafterTime : SequenceTrigger, IValidatable, IDSOTargetProxy
     {
         private IASAOptions options;
         private readonly IMountMediator mountMediator;
@@ -239,6 +239,19 @@ namespace NINA.Photon.Plugin.ASA.MLTP
 
         public override async Task Execute(ISequenceContainer context, IProgress<ApplicationStatus> progress, CancellationToken token)
         {
+            Target = DSOTarget.FindTarget(Parent);
+            if (Target != null)
+            {
+                Logger.Info("Found Target: " + Target);
+                // UpdateChildren(Instructions);
+                Coordinates.Coordinates = Target.InputCoordinates?.Coordinates;
+            }
+            else
+            {
+                Coordinates.Coordinates = telescopeMediator.GetCurrentPosition();
+                Logger.Debug($"MLPTafterTime: Coordinates not set, using telescope coordinates: {Coordinates.Coordinates}");
+            }
+
             var modelBuilderOptions = new ModelBuilderOptions()
             {
                 WestToEastSorting = options.WestToEastSorting,
@@ -255,8 +268,6 @@ namespace NINA.Photon.Plugin.ASA.MLTP
                 ModelPointGenerationType = ModelPointGenerationTypeEnum.SiderealPath
             };
 
-            Coordinates.Coordinates = telescopeMediator.GetCurrentPosition();
-            Logger.Debug($"MLPTafterTime: Coordinates not set, using telescope coordinates: {Coordinates.Coordinates}");
             UpdateModelPoints();
 
             // delete old model
@@ -462,7 +473,7 @@ namespace NINA.Photon.Plugin.ASA.MLTP
                     siderealTrackRADeltaDegrees = value;
 
                     RaisePropertyChanged();
-                    UpdateModelPoints();
+                    //    UpdateModelPoints();
                 }
             }
         }
@@ -637,7 +648,7 @@ namespace NINA.Photon.Plugin.ASA.MLTP
                 StartSeconds = t.Second;
             }
 
-            UpdateModelPoints();
+            //  UpdateModelPoints();
         }
 
         private void UpdateEndTime()
@@ -650,7 +661,7 @@ namespace NINA.Photon.Plugin.ASA.MLTP
                 EndSeconds = t.Second;
             }
 
-            UpdateModelPoints();
+            // UpdateModelPoints();
         }
 
         /*
@@ -775,7 +786,7 @@ namespace NINA.Photon.Plugin.ASA.MLTP
             if (contextCoordinates != null)
             {
                 Coordinates.Coordinates = contextCoordinates.Coordinates;
-                UpdateModelPoints();
+                //   UpdateModelPoints();
                 Inherited = true;
             }
             else
@@ -789,6 +800,26 @@ namespace NINA.Photon.Plugin.ASA.MLTP
         public override string ToString()
         {
             return $"Category: {Category}, Item: {nameof(MLPTafterTime)}, Coordinates: {Coordinates?.Coordinates}, Inherited: {Inherited}, RADelta: {SiderealTrackRADeltaDegrees}, Start: {SelectedSiderealPathStartDateTimeProvider?.Name} ({SiderealTrackStartOffsetMinutes} minutes), Start: {SelectedSiderealPathEndDateTimeProvider?.Name} ({SiderealTrackEndOffsetMinutes} minutes), NumRetries: {BuilderNumRetries}, MaxFailedPoints: {MaxFailedPoints}, MaxPointRMS: {MaxPointRMS}";
+        }
+
+        public InputTarget DSOProxyTarget()
+        {
+            return Target;
+        }
+
+        public InputTarget Target = null;
+
+        public InputTarget FindTarget(ISequenceContainer c)
+        {
+            while (c != null)
+            {
+                if (c is IDSOTargetProxy dso)
+                {
+                    return dso.DSOProxyTarget();
+                }
+                c = c.Parent;
+            }
+            return null;
         }
     }
 }